+++ 9solutions-backend/src/main/java/app/Application.java
@@ -1,4 +1,4 @@
-package CRUD.solutionsbackend;
+package app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
