+++ 9solutions-backend/src/main/java/app/controller/CaixaController.java
@@ -1,10 +1,10 @@
-package CRUD.solutionsbackend;
+package app.controller;
 
+import app.model.Caixa;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
