+++ 9solutions-backend/src/main/java/app/enums/OpcoesDoacao.java
@@ -1,4 +1,4 @@
-package CRUD.solutionsbackend.enums;
+package app.enums;
 
 public class OpcoesDoacao {
     public enum Genero {
