+++ 9solutions-backend/src/main/java/app/model/Usuario.java
@@ -1,4 +1,4 @@
-package CRUD.solutionsbackend;
+package app.model;
 
 public class Usuario {
     private String nomeUser;
